Also refactor the Integrations and Connections sections
