+++ packages/ui/src/ConnectionsGrid.tsx
@@ -1,29 +1,11 @@
 import { CardGrid } from 'patternfly-react';
 import * as React from 'react';
-import { ConnectionSkeleton } from './ConnectionSkeleton';
 
-export interface IConnectionsGridProps {
-  loading: boolean;
-  children: JSX.Element[];
-}
-
-export class ConnectionsGrid extends React.Component<IConnectionsGridProps> {
+export class ConnectionsGrid extends React.Component {
   public render() {
     return (
       <CardGrid fluid={true} matchHeight={true}>
-        <CardGrid.Row>
-          {this.props.loading
-            ? new Array(5).fill(0).map((_, index) => (
-                <CardGrid.Col sm={6} md={3} key={index}>
-                  <ConnectionSkeleton key={index} />
-                </CardGrid.Col>
-              ))
-            : this.props.children.map((c: any, index: number) => (
-                <CardGrid.Col sm={6} md={3} key={index}>
-                  {c}
-                </CardGrid.Col>
-              ))}
-        </CardGrid.Row>
+        <CardGrid.Row>{this.props.children}</CardGrid.Row>
       </CardGrid>
     );
   }
