+++ packages/ui/src/ConnectionsGridCell.tsx
@@ -1,0 +1,12 @@
+import { CardGrid } from 'patternfly-react';
+import * as React from 'react';
+
+export class ConnectionsGridCell extends React.Component {
+  public render() {
+    return (
+      <CardGrid.Col sm={6} md={3}>
+        {this.props.children}
+      </CardGrid.Col>
+    );
+  }
+}
