+++ packages/ui/src/index.ts
@@ -1,12 +1,15 @@
 export * from './AggregatedMetricCard';
 export * from './ConnectionCard';
 export * from './ConnectionsGrid';
+export * from './ConnectionsGridCell';
 export * from './ConnectionSkeleton';
+export * from './ConnectionsListView';
 export * from './Dashboard';
 export * from './IntegrationProgress';
 export * from './IntegrationsList';
 export * from './IntegrationsListItem';
 export * from './IntegrationsListSkeleton';
+export * from './IntegrationsListView';
 export * from './IntegrationStatus';
 export * from './IntegrationStatusDetail';
 export * from './ListViewToolbar';
